dl: checksums: Component class moved here from dlmanager
