--- oc_delivery_apps/dlmanager/manage_dlmanager.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-import os
-import sys
-
-
-if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "dlmanager.settings")
-    from django.core.management import execute_from_command_line
-    execute_from_command_line(sys.argv)
